Extracted sortProductts to utils file
